refactor: extract layout components from client bundles
- Create ListWithFiltersLayout for server-side layout rendering
- Create ListWithFilters for client-side interactive functionality
- Add dynamicSubNav prop for client-side navigation (CastAndCrew)
- Update all list components to remove layout dependencies
- Update all list pages to use server/client component separation
- Eliminates ~81KB ReviewsSubNav chunk from client bundles
- Reduces individual page chunks to 1-7KB each

This architectural change separates static layout components (Backdrop,
Layout, Footer, Mast) from interactive components, keeping them out of
client JavaScript bundles for significant performance improvements.
