+++ src/pages/reviews/index.astro
@@ -1,16 +1,36 @@
 ---
 import { AllReviews } from "~/components/Reviews/AllReviews";
+import { ListWithFiltersLayout } from "~/components/ListWithFiltersLayout";
+import { Backdrop } from "~/components/Backdrop";
+import { ReviewsSubNav } from "~/components/ReviewsSubNav";
 import { getProps } from "~/components/Reviews/getProps";
 import Layout from "~/layouts/Layout.astro";
+import React from "react";
 
 const props = await getProps();
+
+// Separate layout props from interactive props
+const { backdropImageProps, deck, ...interactiveProps } = props;
+
+// Create React elements in frontmatter
+const backdropElement = React.createElement(Backdrop, {
+  deck,
+  imageProps: backdropImageProps,
+  title: "Reviews"
+});
+
+const subNavElement = React.createElement(ReviewsSubNav, {
+  active: "all"
+});
 ---
 
 <Layout
   canonical={true}
   meta={{ description: props.metaDescription }}
-  openGraph={{ description: props.deck, type: "website" }}
+  openGraph={{ description: deck, type: "website" }}
   title="All Reviews"
 >
-  <AllReviews client:load {...props} />
+  <ListWithFiltersLayout backdrop={backdropElement} subNav={subNavElement}>
+    <AllReviews client:load {...interactiveProps} />
+  </ListWithFiltersLayout>
 </Layout>
