+++ src/pages/reviews/overrated/index.astro
@@ -1,9 +1,27 @@
 ---
 import { getPropsForOverrated as getProps } from "~/components/Reviews/getProps";
 import { Overrated } from "~/components/Reviews/Overrated";
+import { ListWithFiltersLayout } from "~/components/ListWithFiltersLayout";
+import { Backdrop } from "~/components/Backdrop";
+import { ReviewsSubNav } from "~/components/ReviewsSubNav";
 import Layout from "~/layouts/Layout.astro";
+import React from "react";
 
 const props = await getProps();
+
+// Separate layout props from interactive props
+const { backdropImageProps, deck, ...interactiveProps } = props;
+
+// Create React elements in frontmatter
+const backdropElement = React.createElement(Backdrop, {
+  deck,
+  imageProps: backdropImageProps,
+  title: "Overrated"
+});
+
+const subNavElement = React.createElement(ReviewsSubNav, {
+  active: "overrated"
+});
 ---
 
 <Layout
@@ -12,10 +30,12 @@ const props = await getProps();
     description: props.metaDescription,
   }}
   openGraph={{
-    description: props.deck,
+    description: deck,
     type: "website",
   }}
   title="Overrated Disappointments"
 >
-  <Overrated client:load {...props} />
+  <ListWithFiltersLayout backdrop={backdropElement} subNav={subNavElement}>
+    <Overrated client:load {...interactiveProps} />
+  </ListWithFiltersLayout>
 </Layout>
