+++ src/pages/viewings/index.astro
@@ -1,16 +1,31 @@
 ---
 import { getProps } from "~/components/Viewings/getProps";
 import { Viewings } from "~/components/Viewings/Viewings";
+import { ListWithFiltersLayout } from "~/components/ListWithFiltersLayout";
+import { Backdrop } from "~/components/Backdrop";
 import Layout from "~/layouts/Layout.astro";
+import React from "react";
 
 const props = await getProps();
+
+// Separate layout props from interactive props
+const { backdropImageProps, deck, ...interactiveProps } = props;
+
+// Create React elements in frontmatter
+const backdropElement = React.createElement(Backdrop, {
+  deck,
+  imageProps: backdropImageProps,
+  title: "Viewing Log"
+});
 ---
 
 <Layout
   canonical={true}
   meta={{ description: props.metaDescription }}
-  openGraph={{ description: props.deck, type: "website" }}
+  openGraph={{ description: deck, type: "website" }}
   title="Viewing Log"
 >
-  <Viewings client:load {...props} />
+  <ListWithFiltersLayout backdrop={backdropElement}>
+    <Viewings client:load {...interactiveProps} />
+  </ListWithFiltersLayout>
 </Layout>
