+++ src/pages/watchlist/index.astro
@@ -1,19 +1,34 @@
 ---
 import { getProps } from "~/components/Watchlist/getProps";
 import { Watchlist } from "~/components/Watchlist/Watchlist";
+import { ListWithFiltersLayout } from "~/components/ListWithFiltersLayout";
+import { Backdrop } from "~/components/Backdrop";
 import Layout from "~/layouts/Layout.astro";
+import React from "react";
 
 const props = await getProps();
+
+// Separate layout props from interactive props
+const { backdropImageProps, deck, ...interactiveProps } = props;
+
+// Create React elements in frontmatter
+const backdropElement = React.createElement(Backdrop, {
+  deck,
+  imageProps: backdropImageProps,
+  title: "Watchlist"
+});
 ---
 
 <Layout
   canonical={true}
   meta={{ description: props.metaDescription }}
   openGraph={{
-    description: props.deck,
+    description: deck,
     type: "website",
   }}
   title="Watchlist"
 >
-  <Watchlist client:load {...props} />
+  <ListWithFiltersLayout backdrop={backdropElement}>
+    <Watchlist client:load {...interactiveProps} />
+  </ListWithFiltersLayout>
 </Layout>
